+++ assets/scss/theme/index.scss
@@ -1,6 +1,7 @@
 body {
   background: #F1F3F6;
   font-family: 'Montserrat', sans-serif;
+  cursor: default;
   a {
     text-decoration: none!important;
   }
@@ -72,15 +73,18 @@ body {
   width: 100vw;
   height: 100vh;
   .sidebar{
-    width: 20%;
+    width: 20vw;
     height: 100%;
     background-color: white;
     position: fixed;
+    overflow-y: scroll;
     padding-left: 50px;
+    margin-bottom: 10px;
     &--logo {
       margin-bottom: 20px;
     }
     &--navigation {
+      margin-top: 60px;
       padding-top: 5px;
       padding-bottom: 5px;
       &--container:hover {
@@ -116,7 +120,36 @@ body {
         }
       }
     }
+    &--sidebarquick{
+      margin-top: 14vh;
+      margin-right: 80px;
+      padding-left: 25px;
+      padding-right: 25px;
+      padding-top: 9px;
+      height: 200px;
+      width: 200px;
+      background: #F5F5FB;
+      border-radius: 12px;
+      .quick-box {
+        font-size: 16px;
+        line-height: 22px;
+        letter-spacing: 0em;
+        text-align: center;
+        color: rgba(34, 33, 91, 0.6);
+        margin-top: 36px;
+      }
+      .quick-button {
+        margin-top: 42px;
+      }
+    }
   }
+}
+
+
+.page_label {
+  font-size: 20px;
+  font-weight: 500;
+  line-height: 23.44px
 }
 
 .navbar {
@@ -142,19 +175,34 @@ body {
 }
 
 .appnnav {
-  width: 80%;
-  height: 13vh;
-  margin-left: 20%;
-  position: fixed;
+
 }
 
+
+
 .mainscreen {
-  width: 80%;
-  margin-top: 13vh;
-  height: 80vh;
-  padding-left: 10px;
-  padding-right: 10px;
+  width: 80vw;
+  height: 100vh;
+  // padding: 10px;
   margin-left: 20%;
+  height: 100vh;
+  /* margin-right: 10px; */
   position: fixed;
   color: #3F2668;
+  overflow-y: scroll;
+  .maincontent{
+    margin-top: 2vh;
+    margin-left: 10px;
+    margin-right: 10px;
+    height: 100%;
+    .container {
+      margin: 0px;
+    }
+  }
+}
+
+@media (min-width: 1200px) {
+  .container, .container-sm, .container-md, .container-lg, .container-xl {
+    max-width: 1195px;
+  }
 }
